--- emails/sesclient.go
@@ -1,21 +1,0 @@
-package emails
-
-import (
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/ses"
-)
-
-func NewSESClient(key, secret, region string) (*ses.SES, error) {
-	sess, err := session.NewSession(&aws.Config{
-		Region:      aws.String(region),
-		Credentials: credentials.NewStaticCredentials(key, secret, ""),
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return ses.New(sess), nil
-}
